remove deprecated ParseSubjectStringToRawDERBytes function & refactor and move tests
