+++ pkg/util/pki/subject.go
@@ -126,13 +126,3 @@ func ExtractCommonNameFromRDNSequence(rdns pkix.RDNSequence) string {
 
 	return ""
 }
-
-// Deprecated: this function will be removed in a future release.
-func ParseSubjectStringToRawDERBytes(subject string) ([]byte, error) {
-	rdnSequence, err := UnmarshalSubjectStringToRDNSequence(subject)
-	if err != nil {
-		return nil, err
-	}
-
-	return MarshalRDNSequenceToRawDERBytes(rdnSequence)
-}
